XML parser. Platform. main() cleanup.
Added new functions to XML parser - it can now retrieve attribute values from nested nodes as strings as well.
Created a GameObject abstract class which will be the basis for elements that use sprites, such as Platform and Bricks.
Created the Platform class and moved the code from prototype. The platform gets a texture from an xml file and moves based on mouse position.