Ball movement. Movement refactor.
Implemented ball movement: at the start of the game the ball is moving based on mouse position, similar to platform. The game starts on left mouse click and the ball bounces off the platform.
 !!! Should probably solve game start with observer instead of setting a bool. !!!
Refactored platform movement and ball movement to use delta time for frame rate independency and setting the position with sprite.move().