States - MainMenuState.
Implemented a main menu with a title and two buttons.
On Start button click, Game switches state to PlayingState.
Moved UIManager to PlayingState since it's not necessary for any other state. It's no longer observing PlayingState via Game class, but directly.
Moved the logic for text creation from UIManager to a separate class so it can be used by other classes as well.