Bug fix in BrickGrid. Empty spaces.
Fixed a bug where game crashed if no impenetrable bricks were present in the brick layout by introducing another member to GridData (bool canDestroy).
Fixed the handling of empty spaces - instead of instantiating an impenetrable brick which doesn't render, now empty spaces are literally empty GridData and are destroyed when  the rest of the bricks get returned to the pool at the level end.
Started working on loading the values for a next level (GameConfig event).