Collision System
After weeks of breaking the game and reassembling it, it's finally starting to work the way I wanted it to.
CollisionManager will work with Collidable components of objects that can collide. Currently, the ball and the platform get successfully registered for collision. Next, I will try to make the manager actually detect collision. Also, I need to extract the collision logic from Ball.