Collision System: Ball Reflection and Collision Observers
Currently all logic for ball collision and consequent reflection off of collided objects is still in Ball. This should be moved to BallMovement and/or BallCollision.
BrickGridVisual SHOULD NOT handle anything collision related, and it currently does --> move either to Brick class or BrickGrid.
PROBLEM: BrickGrid's way of handling collisions.
Also consider clearing the collision observers' vector in Collision Manager and registering observers again at the start of each PlayingState.