Collision System complete. Decoupling Ball & Platform.
Collision system is now functional.
TO DO: refactor the Ball class and remove ball collision to a separate class.

Platform is no longer needed to instantiate a Ball. Instead I introduced a class called GameObjectData which only takes the platform's sprite and gives it to Ball if the Ball requests it from LevelDataProvider.
TO DO: try to ensure there is platform data before another class tries to use it (sort of like lazy initialiization).